Organize resolvers into a new folder
